Component API for the Multi-Page Review Screen
Issue #172
