+++ ginivision/src/main/java/net/gini/android/vision/review/multipage/MultiPageReviewFragmentInterface.java
@@ -1,0 +1,17 @@
+package net.gini.android.vision.review.multipage;
+
+import android.support.annotation.NonNull;
+
+import net.gini.android.vision.document.GiniVisionMultiPageDocument;
+
+/**
+ * Created by Alpar Szotyori on 07.05.2018.
+ *
+ * Copyright (c) 2018 Gini GmbH.
+ */
+public interface MultiPageReviewFragmentInterface {
+
+    GiniVisionMultiPageDocument getMultiPageDocument();
+
+    void setListener(@NonNull final MultiPageReviewFragmentListener listener);
+}
