+++ ginivision/src/main/java/net/gini/android/vision/review/multipage/previews/PreviewsPageChangeListener.java
@@ -1,0 +1,11 @@
+package net.gini.android.vision.review.multipage.previews;
+
+/**
+ * Created by Alpar Szotyori on 08.05.2018.
+ *
+ * Copyright (c) 2018 Gini GmbH.
+ */
+public interface PreviewsPageChangeListener {
+
+    void onPageSelected(final int position);
+}
