+++ ginivision/src/main/java/net/gini/android/vision/review/multipage/thumbnails/ThumbnailChangeListener.java
@@ -1,0 +1,13 @@
+package net.gini.android.vision.review.multipage.thumbnails;
+
+/**
+ * Created by Alpar Szotyori on 08.05.2018.
+ *
+ * Copyright (c) 2018 Gini GmbH.
+ */
+public interface ThumbnailChangeListener {
+
+    void onThumbnailMoved();
+
+    void onThumbnailSelected(final int position);
+}
