+++ ginivision/src/main/java/net/gini/android/vision/review/multipage/thumbnails/ThumbnailsTouchHelperCallback.java
@@ -1,0 +1,53 @@
+package net.gini.android.vision.review.multipage.thumbnails;
+
+import android.support.v7.widget.RecyclerView;
+import android.support.v7.widget.helper.ItemTouchHelper;
+
+/**
+ * Created by Alpar Szotyori on 08.05.2018.
+ *
+ * Copyright (c) 2018 Gini GmbH.
+ */
+public class ThumbnailsTouchHelperCallback extends ItemTouchHelper.Callback {
+
+    private final ThumbnailsTouchHelperListener mListener;
+
+    public ThumbnailsTouchHelperCallback(
+            final ThumbnailsTouchHelperListener listener) {
+        mListener = listener;
+    }
+
+    @Override
+    public int getMovementFlags(final RecyclerView recyclerView,
+            final RecyclerView.ViewHolder viewHolder) {
+        return makeMovementFlags(ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT,
+                0);
+    }
+
+    @Override
+    public boolean onMove(final RecyclerView recyclerView,
+            final RecyclerView.ViewHolder viewHolder,
+            final RecyclerView.ViewHolder target) {
+        mListener.onItemMove(viewHolder, viewHolder.getAdapterPosition(), target,
+                target.getAdapterPosition());
+        return true;
+    }
+
+    @Override
+    public boolean isItemViewSwipeEnabled() {
+        return false;
+    }
+
+    @Override
+    public void onSwiped(final RecyclerView.ViewHolder viewHolder, final int direction) {
+    }
+
+    @Override
+    public void onSelectedChanged(final RecyclerView.ViewHolder viewHolder,
+            final int actionState) {
+        super.onSelectedChanged(viewHolder, actionState);
+        if (actionState == ItemTouchHelper.ACTION_STATE_IDLE) {
+            mListener.onDragFinished();
+        }
+    }
+}
