+++ ginivision/src/main/res/layout/gv_fragment_multi_page_review.xml
@@ -1,0 +1,115 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/root_view"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@android:color/black"
+    tools:context=".review.multipage.MultiPageReviewActivity">
+
+    <android.support.v4.view.ViewPager
+        android:id="@+id/gv_view_pager"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_above="@id/gv_thumbnails_panel" />
+
+    <TextView
+        android:id="@+id/gv_page_indicator"
+        style="@style/TextAppearance.AppCompat.Medium"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_above="@id/gv_thumbnails_panel"
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="16dp"
+        android:background="@drawable/gv_page_indicator_background"
+        android:gravity="center"
+        android:paddingBottom="4dp"
+        android:paddingEnd="8dp"
+        android:paddingLeft="8dp"
+        android:paddingRight="8dp"
+        android:paddingStart="8dp"
+        android:paddingTop="4dp"
+        android:textColor="@android:color/white"
+        tools:text="2 von 4" />
+
+    <ImageButton
+        android:id="@+id/gv_button_next"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_above="@+id/gv_thumbnails_panel"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentRight="true"
+        android:layout_marginBottom="@dimen/gv_fab_vertical_margin"
+        android:layout_marginEnd="@dimen/gv_fab_horizontal_margin"
+        android:layout_marginRight="@dimen/gv_fab_horizontal_margin"
+        android:background="@drawable/gv_review_fab_background"
+        android:src="@drawable/gv_review_fab_checkmark" />
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/gv_thumbnails_panel"
+        android:layout_width="match_parent"
+        android:layout_height="180dp"
+        android:layout_alignTop="@id/gv_toolbar"
+        android:background="#262626"
+        android:clipToPadding="false"
+        android:paddingEnd="8dp"
+        android:paddingLeft="8dp"
+        android:paddingRight="8dp"
+        android:paddingStart="8dp" />
+
+    <LinearLayout
+        android:id="@+id/gv_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="48dp"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true"
+        android:background="#262626"
+        android:gravity="center"
+        android:orientation="horizontal">
+
+        <android.support.v4.widget.Space
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            android:layout_weight="1" />
+
+        <ImageButton
+            android:id="@+id/gv_button_rotate"
+            style="@style/Widget.AppCompat.Button.Borderless"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:background="@drawable/gv_toolbar_button_background"
+            app:srcCompat="@drawable/gv_rotate_icon" />
+
+        <android.support.v4.widget.Space
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            android:layout_weight="2" />
+
+        <ImageButton
+            android:id="@+id/gv_button_reorder"
+            style="@style/Widget.AppCompat.Button.Borderless"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:background="@drawable/gv_toolbar_button_background"
+            app:srcCompat="@drawable/gv_reorder_icon" />
+
+        <android.support.v4.widget.Space
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            android:layout_weight="2" />
+
+        <ImageButton
+            android:id="@+id/gv_button_delete"
+            style="@style/Widget.AppCompat.Button.Borderless"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            android:background="@drawable/gv_toolbar_button_background"
+            app:srcCompat="@drawable/gv_delete_icon" />
+
+        <android.support.v4.widget.Space
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            android:layout_weight="1" />
+    </LinearLayout>
+</RelativeLayout>
