+++ ginivision/src/main/res/layout/gv_item_multi_page_preview.xml
